chore: use native js modules
