+++ index.html
@@ -83,7 +83,7 @@
           </label>
           <label>
             Hide Checked
-            <input type="checkbox" id="autoSort" class="checkbox" />
+            <input type="checkbox" id="hideChecked" class="checkbox" />
           </label>
           <!-- TODO: Sort by Section -->
           <!-- <label>
@@ -102,6 +102,11 @@
         <p id="priceAmount"></p>
       </div>
     </div>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
+    <script type="module" src="./modules/db.js"></script>
+    <script type="module" src="./modules/domElements.js"></script>
+    <script type="module" src="./modules/populateItems.js"></script>
+    <script type="module" src="./modules/form.js"></script>
+    <script type="module" src="./modules/itemsControl.js"></script>
   </body>
 </html>
