+++ modules/domElements.js
@@ -1,0 +1,16 @@
+// get the UI elements we need
+export const itemForm = document.getElementById('itemForm');
+export const itemsDiv = document.getElementById('itemsDiv');
+export const priceAmount = document.getElementById('priceAmount');
+export const clearFormButton = document.getElementById('cancelButton');
+// form inputs and err msgs
+export const nameInput = document.getElementById('nameInput');
+export const nameError = document.getElementById('nameError');
+export const priceInput = document.getElementById('priceInput');
+export const priceError = document.getElementById('priceError');
+// show/hide filter ui
+export const showItemsControlContainer =
+  document.getElementById('showItemsControl');
+export const itemsControl = document.getElementById('itemsControl');
+export const toggleMessage = document.getElementById('toggleMessage');
+export const toggleArrow = document.getElementById('arrow');
